+++ api/tests/evaluation/integration/infrastructure/repositories/certification-completed-job-repository_test.js
@@ -1,5 +1,5 @@
-import { certificationCompletedJobRepository } from '../../../../../lib/infrastructure/repositories/jobs/certification-completed-job-repository.js';
 import { CertificationCompletedJob } from '../../../../../src/certification/evaluation/domain/events/CertificationCompleted.js';
+import { certificationCompletedJobRepository } from '../../../../../src/evaluation/infrastructure/repositories/certification-completed-job-repository.js';
 import { LOCALE } from '../../../../../src/shared/domain/constants.js';
 import { JobPriority } from '../../../../../src/shared/infrastructure/repositories/jobs/job-repository.js';
 import { expect } from '../../../../test-helper.js';
